Create README.md
This project is used to test Using classes by creating separate files in order to provide flexibility.  The main Class called the PlayerLinkedListTest is used to Test DataStructure, Player, PlayerLinkedList, PlayerNode and ShellLinkedList Classes.
In the PlayerLinkeListTest 3 objects of Player is created to input 3 different values, the ID, Persons name and Game name.
It shall display The ID, Name and Game name as well as deleting files from the linked list classes using a temp object of Player.
This project is running Java Eclipse.